+++ src/classes/simple-calendar-notes.ts
@@ -71,7 +71,7 @@ export class SimpleCalendarNotes extends FormApplication {
         const options = super.defaultOptions;
         options.template = "modules/foundryvtt-simple-calendar/templates/calendar-notes.html";
         options.title = "FSC.Notes.DialogTitle";
-        options.classes = ["form","simple-calendar-note"];
+        options.classes = ["form","simple-calendar simple-calendar-note"];
         options.resizable = true;
         options.closeOnSubmit = false;
         options.width = 500;